requirements.txt
faiss-cpu
sentence-transformers
streamlit
pypdf
python-dotenv
numpy
scikit-learn
requests
